# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: Full Revenue Loan API
  version: "1.0.0"
  description: |
    Backend API for the Rappi Full Revenue Loan Option feature.

    Set `DEMO_MODE=true` to use stub data and skip real Syntage/Google calls.
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://full-revenue-backend-HASH-uc.a.run.app
    description: Production (Cloud Run) — replace HASH with actual service hash
...
paths:
...
  /full-revenue/applications:
    post:
      tags: [Applications]
      summary: Create a new loan application
      description: |
        Creates an application record with status UNDERWRITING_PENDING.
        Returns the application ID used in all subsequent calls.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
            example:
              merchant_id: "rappi-merchant-12345"
      responses:
        "201":
          description: Application created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /full-revenue/applications/{id}:
    get:
      tags: [Applications]
      summary: Get application by ID
      description: Returns full application data. Poll this endpoint every 5 seconds to check underwriting status.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Application data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  schemas:
    DecisionStatus:
      type: string
      enum:
        - UNDERWRITING_PENDING
        - APPROVED
        - REJECTED
        - MANUAL_REVIEW
      description: |
        - UNDERWRITING_PENDING: Application submitted, evaluation in progress
        - APPROVED: Revenue exceeds threshold, loan approved
        - REJECTED: Revenue below threshold
        - MANUAL_REVIEW: Insufficient data or edge case requiring human review
    CreateApplicationRequest:
      type: object
      required: [merchant_id]
      properties:
        merchant_id:
          type: string
          minLength: 1
          example: rappi-merchant-12345
    FormData:
      type: object
      required:
        - legal_name
        - tax_id
        - address
        - bank_account
        - email
        - monthly_revenue_estimate
        - revenue_sources
        - consent_given
        - google_connected
      properties:
        legal_name:
          type: string
          minLength: 2
          maxLength: 200
          example: "Restaurante El Buen Sabor S.A. de C.V."
        tax_id:
          type: string
          minLength: 8
          maxLength: 20
          example: "REST123456ABC"
        address:
          type: string
          minLength: 5
          maxLength: 500
          example: "Av. Insurgentes Sur 1234, Col. Del Valle, CDMX"
        bank_account:
          type: string
          minLength: 10
          maxLength: 30
          example: "012345678901234567"
        email:
          type: string
          format: email
          example: "contacto@restaurante.com"
        monthly_revenue_estimate:
          type: number
          minimum: 0.01
          example: 85000
        revenue_sources:
          type: array
          minItems: 1
          items:
            type: string
            enum:
              - ventas_rappi
              - ventas_propias
              - exportaciones
              - servicios
              - otros
          example: ["ventas_rappi", "ventas_propias"]
        notes:
          type: string
          maxLength: 1000
          nullable: true
          example: "Tenemos un catering adicional los fines de semana"
        consent_given:
          type: boolean
          enum: [true]
          description: Must be exactly `true`. Any other value is rejected.
        google_connected:
          type: boolean
          description: Whether the user completed the Google OAuth flow
...
    DecisionPayload:
      type: object
      properties:
        reason:
          type: string
        total_revenue:
          type: number
          description: Monthly revenue used for decision (max of available sources)
        threshold_used:
          type: number
          description: APPROVAL_THRESHOLD value at time of decision
        data_sources:
          type: array
          items:
            type: string
            enum: [syntage, google]
        decided_at:
          type: string
          format: date-time
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
        decision_status:
          $ref: "#/components/schemas/DecisionStatus"
        form_data:
          $ref: "#/components/schemas/FormData"
          nullable: true
        decision_payload:
          $ref: "#/components/schemas/DecisionPayload"
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Validation failed
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Application not found